requirements.txt
Fabric==2.5.0
patchwork==1.0.1
PyYAML==5.1.2
python-debian==0.1.35
chardet==3.0.4
requests==2.22.0
coverage==4.5.4
nose==1.3.7
tox==2.7.0
docker==4.0.2
docker-pycreds==0.4.0
tox-docker==1.4.1
pytest==5.1.1
